bugfixes, auto-push

+++ ccb/worktree.py
@@ -5,30 +5,39 @@ from .recipe import Recipe
 
 
 class RecipeInWorktree:
-    def __init__(self, recipe, branch_name):
+    def __init__(self, recipe):
         self.recipe = recipe
-        self.branch_name = branch_name
+        self.tmpdir = None
 
     def __enter__(self):
         self.tmpdir = tempfile.TemporaryDirectory()
-        subprocess.check_call(
-            [
-                "git",
-                "worktree",
-                "add",
-                "-q",
-                "--checkout",
-                "-b",
-                self.branch_name,
-                self.tmpdir.name,
-                "master",
-            ],
-            cwd=self.recipe.path,
-        )
-        return Recipe(self.tmpdir.name, self.recipe.name)
+        try:
+            subprocess.check_call(
+                [
+                    "git",
+                    "worktree",
+                    "add",
+                    "-q",
+                    "--checkout",
+                    "--detach",
+                    self.tmpdir.name,
+                    "master",
+                ],
+                cwd=self.recipe.path,
+            )
+            return Recipe(self.tmpdir.name, self.recipe.name)
+        except:
+            self.cleanup()
+            raise
 
     def __exit__(self, exc_type, exc_value, exc_traceback):
-        subprocess.check_output(
+        self.cleanup()
+
+    def cleanup(self):
+        if self.tmpdir is None:
+            return
+
+        subprocess.call(
             ["git", "worktree", "remove", "-f", self.tmpdir.name], cwd=self.recipe.path
         )
         self.tmpdir.cleanup()
